Extract ZendeskTicketType to its own file
